refactor; move tests to Makefile
